Only include ipmid/api.hpp for the new API
After some feedback from users of the new IPMI API, they wanted to see
two things:
 1) don't require ipmid/api.hpp and ipmid/registration.hpp to be able to
    write new handlers
 2) only require including ipmid/api.hpp (instead of ipmid/api.h)

So now, by simply including ipmid/api.hpp instead of ipmid/api.h
(deprecated), handlers incorporating the new IPMI API can be written.
