+++ app/watchdog_service.cpp
@@ -1,8 +1,7 @@
 #include "watchdog_service.hpp"
 
-#include <ipmid/api.h>
-
 #include <exception>
+#include <ipmid/api.hpp>
 #include <phosphor-logging/elog-errors.hpp>
 #include <phosphor-logging/elog.hpp>
 #include <phosphor-logging/log.hpp>
