+++ groupext.cpp
@@ -1,6 +1,5 @@
-#include <ipmid/api.h>
-
 #include <cstdio>
+#include <ipmid/api.hpp>
 
 #define GRPEXT_GET_GROUP_CMD 0
 void register_netfn_groupext_functions() __attribute__((constructor));
