+++ ipmid-new.cpp
@@ -31,7 +31,6 @@
 #include <ipmid/handler.hpp>
 #include <ipmid/message.hpp>
 #include <ipmid/oemrouter.hpp>
-#include <ipmid/registration.hpp>
 #include <ipmid/types.hpp>
 #include <map>
 #include <memory>
