+++ sensordatahandler.hpp
@@ -2,9 +2,8 @@
 
 #include "sensorhandler.hpp"
 
-#include <ipmid/api.h>
-
 #include <cmath>
+#include <ipmid/api.hpp>
 #include <ipmid/types.hpp>
 #include <ipmid/utils.hpp>
 #include <sdbusplus/message/types.hpp>
